renaming module for clarity, also use the new `create_system` function
